% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraceptives.r
\docType{data}
\name{contraceptives}
\alias{contraceptives}
\title{Oral contraceptive use and breast cancer rates}
\format{A 3-dimensional table.

\enumerate{
  \item \code{OC Usage}: Subject exposure to oral contraceptives.
  \item \code{Disease Status}: Breast cancer present (case) or absent (control).
  \item \code{Age Group}: Age group of the subject.

}}
\source{
Hennekens, C. H., F. E. Speizer, R. J. Lipnick, B. Rosner, C. Bain,
C. Belanger, M. J. Stampfer, W. Willett, and R. Peto. 1984. "A Case-Control
Study of Oral Contraceptive Use and Breast Cancer." \emph{Journal of the
National Cancer Institute} \strong{72} (1): 39–42. Table 1.
}
\usage{
data(contraceptives)
}
\description{
This data summrizes the raw counts of the case-control study investigating
the link between breast cancer rates and oral contraceptive use, stratified
by age group. In toto, 10,890 subjects. See source for details.
}
\keyword{datasets}
