% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.cmh.test.r
\name{print.power.cmh}
\alias{print.power.cmh}
\title{Print \code{"power.cmh"} object}
\usage{
\method{print}{power.cmh}(x, detail = TRUE, n.frac = FALSE, ...)
}
\arguments{
\item{x}{A \code{"power.cmh"} object.}

\item{detail}{Logical to toggle detailed or simple output.}

\item{n.frac}{Logical indicating whether sample n's should be rounded to the
next whole number.}

\item{...}{Ignored.}
}
\description{
The S3 print method for the \code{"power.cmh"} object
}
