% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.risk.r
\name{rel.risk}
\alias{rel.risk}
\title{Calculate the relative risk from a 2-by-2 table}
\usage{
rel.risk(x, col.num = 1)
}
\arguments{
\item{x}{A table or matrix containing frequencies}

\item{col.num}{The column number upon which relative risk should be calculated}
}
\value{
A numeric vector.
}
\description{
Computes the relative risk of a specified column of a two-by-two table.
}
\author{
Paul W. Egeler, M.S.
}
